+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 NOTE: Subcrates have their own changelogs: [bevy-tnua-physics-integration-layer](physics-integration-layer/CHANGELOG.md), [bevy-tnua-rapier](rapier3d/CHANGELOG.md), [bevy-tnua-avian](avian3d/CHANGELOG.md).
 
 ## [Unreleased]
+
+## 0.25.0 - 2025-09-26
 ### Removed
 - [**BREAKING**] `Default` implementation from `TnuaControllerPlugin` and
   `TnuaCrouchEnforcer`. Since the schedule must match the physics backend